refactor: 可読性を向上
