--- tools/util.ts
@@ -1,21 +1,0 @@
-import axios from 'axios'
-
-interface IImasAPIResponse {
-  results: {
-    bindings: any[]
-  }
-}
-
-export async function fetchIdolData(query: string) {
-  const url = `https://sparql.crssnky.xyz/spql/imas/query?output=json&query=${encodeURIComponent(
-    query
-  )}`
-
-  try {
-    const res = await axios.get<IImasAPIResponse>(url, { timeout: 5000 })
-    return res.data.results.bindings
-  } catch (err) {
-    console.error(err)
-    throw err
-  }
-}
